Match file name to class name
