+++ lib/bitmovin_player.dart
@@ -3,8 +3,8 @@ export 'src/api/analytics/analytics_config.dart';
 export 'src/api/analytics/custom_data.dart';
 export 'src/api/analytics/default_metadata.dart';
 export 'src/api/analytics/source_metadata.dart';
-export 'src/api/casting/casting_api.dart';
-export 'src/api/casting/casting_config.dart';
+export 'src/api/casting/bitmovin_cast_manager_api.dart';
+export 'src/api/casting/bitmovin_cast_manager_options.dart';
 export 'src/api/drm/drm_config.dart';
 export 'src/api/drm/fairplay_config.dart';
 export 'src/api/drm/widevine_config.dart';
